+++ cs8803/Services/FriendshipService.swift
@@ -5,7 +5,6 @@
 //  Created by Ethan Yan on 21/1/25.
 //
 
-
 import FirebaseFirestore
 import FirebaseAuth
 
@@ -19,19 +18,16 @@ class FriendshipService {
         let userADocRef = db.collection("users").document(userA)
         let userBDocRef = db.collection("users").document(userB)
 
-        // References for each subcollection doc
         let docRefInB = userBDocRef.collection("friends").document(userA)
         let docRefInA = userADocRef.collection("friends").document(userB)
 
         db.runTransaction({ transaction, errorPointer in
-            // 1) Create or update doc in B’s subcollection => pending
             transaction.setData([
                 "status": "pending",
                 "requestSender": userA,
                 "lastUpdated": FieldValue.serverTimestamp()
             ], forDocument: docRefInB, merge: true)
 
-            // 2) Create or update doc in A’s subcollection => pending
             transaction.setData([
                 "status": "pending",
                 "requestSender": userA,
@@ -44,7 +40,6 @@ class FriendshipService {
         }
     }
 
-    
     // MARK: - Accept Friend Request
     func acceptFriendRequest(from userA: String, completion: @escaping (Error?) -> Void) {
         guard let userB = Auth.auth().currentUser?.uid else { return }
@@ -52,19 +47,15 @@ class FriendshipService {
         let userBDocRef = db.collection("users").document(userB)
         let userADocRef = db.collection("users").document(userA)
         
-        // 1) Update B's record about A => accepted
         let docRefB = userBDocRef.collection("friends").document(userA)
-        // 2) Also create A's record about B => accepted
         let docRefA = userADocRef.collection("friends").document(userB)
         
         db.runTransaction({ transaction, errorPointer in
-            // B's doc => accepted
             transaction.setData([
                 "status": "accepted",
                 "lastUpdated": FieldValue.serverTimestamp()
             ], forDocument: docRefB, merge: true)
             
-            // A's doc => accepted
             transaction.setData([
                 "status": "accepted",
                 "lastUpdated": FieldValue.serverTimestamp()
@@ -75,8 +66,8 @@ class FriendshipService {
             completion(error)
         }
     }
-    
-    // MARK: - Reject or Remove Friend
+
+    // MARK: - Remove Friend or Reject Request
     func removeFriend(userID: String, completion: @escaping (Error?) -> Void) {
         guard let currentUID = Auth.auth().currentUser?.uid else { return }
 
@@ -89,26 +80,54 @@ class FriendshipService {
             completion(error)
         }
     }
-    
-    // MARK: - Fetch My Friends
-    func fetchAcceptedFriends(completion: @escaping ([String]) -> Void) {
+
+    // MARK: - Fetch Friends (Both Accepted & Pending)
+    func fetchFriends(completion: @escaping ([String: String]) -> Void) {
         guard let currentUID = Auth.auth().currentUser?.uid else {
-            completion([])
+            completion([:])
             return
         }
         
         db.collection("users")
             .document(currentUID)
             .collection("friends")
-            .whereField("status", isEqualTo: "accepted")
             .getDocuments { snapshot, error in
                 if let error = error {
                     print("Error fetching friends: \(error.localizedDescription)")
+                    completion([:])
+                    return
+                }
+                
+                var friendList: [String: String] = [:]  // userID -> status
+                snapshot?.documents.forEach { doc in
+                    let data = doc.data()
+                    let status = data["status"] as? String ?? "none"
+                    friendList[doc.documentID] = status
+                }
+                completion(friendList)
+            }
+    }
+
+    // MARK: - Fetch Only Pending Requests
+    func fetchPendingRequests(completion: @escaping ([String]) -> Void) {
+        guard let currentUID = Auth.auth().currentUser?.uid else {
+            completion([])
+            return
+        }
+
+        db.collection("users")
+            .document(currentUID)
+            .collection("friends")
+            .whereField("status", isEqualTo: "pending")
+            .getDocuments { snapshot, error in
+                if let error = error {
+                    print("Error fetching pending requests: \(error.localizedDescription)")
                     completion([])
                     return
                 }
-                let friendIDs = snapshot?.documents.map { $0.documentID } ?? []
-                completion(friendIDs)
+                
+                let pendingUsers = snapshot?.documents.map { $0.documentID } ?? []
+                completion(pendingUsers)
             }
     }
 }
